Created a test factory to support multipe test types.
